// app/enums.ts
export enum DegreePath {
    Undergraduate = 'undergraduate',
    Graduate = 'graduate',
    GraduateCertificate = 'graduateCertificate',
    None = 'none'
}

export enum UndergradDegree {
    BaArtsDigitalComm = 'bachelorOfArtsInDigitalCommunication',
    BaSciSimAndGameDesign = 'bachelorOfScienceInSimulationAndGameDesign'
}

export enum GradDegree {
    MaArtsIntegDesign = 'masterOfArtsInIntegratedDesign',
    MaFineArtsIntegDesign = 'masterOfFineArtsInIntegratedDesign',
    MaSciInterDesignAndInfoArch = 'masterOfScienceInInteractionDesignAndInformationArchitecture',
    DocSciInfoAndInterDesign = 'doctorOfScienceInInformationAndInteractionDesign'
}

export enum GradCertDegree {
    CertDigitialComm = 'certificateInDigitalCommunication',
    CertUserExpDesign = 'certificateInUserExperienceDesign'
}